[TASK] Allow POST method, payload data for POST and headers in endpoints config
To make it possible to protect the health check endpoints,
the monitor must be able to make POST requests, allow to
define headers to send along and also eventually additional
request payload data.

To accomplish that, the csv endpoint configuration is replaced
with a php array for the simplicity and related code is adopted
to deal with the slightly enhanced endpoint configuration more
directly which also allows us to drop an obselete class now.

Let's keep that simple for now.